refactor(js): move generators from js/client to js/js
