--- packages/js/src/index.test.ts
@@ -1,9 +1,0 @@
-// Temporary content, for `pnpm test` to pass CI, will be deleted later
-import { TODO } from '.';
-
-describe('TODO', () => {
-  test('TODO', async () => {
-    const todo: TODO = 'todo';
-    expect(todo).toEqual(todo);
-  });
-});
