+++ packages/js/src/index.ts
@@ -1,4 +1,2 @@
-// Temporary content, for `pnpm build` to pass CI, will be deleted later
-import { Optional } from '@silverhand/essentials';
-
-export type TODO = Optional<string>;
+export * from './constants';
+export * from './utils/generators';
